backend/api/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf_yasg==1.21.5
requests==2.28.2
